Server GUI
New Server GUI
when you start the server, a GUI window will appear, it will have a button to start and stop your media server easily, show logs, and a button to take you straight to your servers site. closing the application will minimize it into the system tray, where it can be easily managed as well.

-----------------Bug Fixes------------------------
fixed a bug that prevented most drives from appearing in the fsbrowser